+++ service/register_instructions.go
@@ -32,18 +32,23 @@ var Ins Instructions
 // 注册指令，绑定文本指令对应的处理方法
 func RegisterInstructions() {
 	Ins = NewInstructions()
+	// 主功能
 	Ins.Bind(UpdateData, "更新数据", "更新")
 	Ins.Bind(HelpGuide, "帮助", "说明")
 	Ins.Bind(Feedback, "反馈", "建议")
-	Ins.Bind(GalleryWebsite, "图鉴网", "图鉴")
-	Ins.Bind(TermInfo, "游戏术语", "术语", "黑话")
 	Ins.Bind(ChefQuery, "厨师", "厨子")
 	Ins.Bind(EquipmentQuery, "厨具", "装备")
 	Ins.Bind(RecipeQuery, "菜谱")
-	Ins.Bind(GuestQuery, "贵客")
+	Ins.Bind(GuestQuery, "贵客", "贵宾")
 	Ins.Bind(AntiqueQuery, "符文")
 	Ins.Bind(CondimentQuery, "调料")
 	Ins.Bind(QuestQuery, "任务", "主线", "支线")
 	Ins.Bind(UpgradeGuestQuery, "碰瓷", "升阶贵客")
 	Ins.Bind(ComboQuery, "后厨", "合成")
+
+	// 辅助信息
+	Ins.Bind(GalleryWebsite, "图鉴网", "图鉴")
+	Ins.Bind(TermInfo, "游戏术语", "术语", "黑话")
+	Ins.Bind(BcjhAppDownload, "白菜菊花")
+	Ins.Bind(BanRandomGif, "涩图", "色图")
 }
